Adjust to updated API for filter params/buffers.
Previously the render_metadata was a struct instead of a function and the param types were a couple of fields (now they are a sub struct) and buffers were structured rather differently. This version was tested on the Mac (64-bit) with the 2018/02/15 binary release of Halide.

Note this change does not address the issue that the binary release copy of libHalide.dylib (instead of libHalide.so) has an absolute library load path. This change only addresses the API changes.